+++ tests/Fixture/Hello/Errors/SemanticObject/DuplicateModelModule.php
@@ -65,4 +65,24 @@ class DuplicateModelModule extends AbstractUi5Module
     {
         return [];
     }
+
+    public function getDashboards(): array
+    {
+        return [];
+    }
+
+    public function getReports(): array
+    {
+        return [];
+    }
+
+    public function getDialogs(): array
+    {
+        return [];
+    }
+
+    public function getName(): string
+    {
+        return 'DuplicateModel';
+    }
 }