+++ tests/Fixture/Hello/Errors/Settings/DuplicateSettingModule.php
@@ -67,4 +67,24 @@ class DuplicateSettingModule extends AbstractUi5Module
     {
         return [];
     }
+
+    public function getDashboards(): array
+    {
+        return [];
+    }
+
+    public function getReports(): array
+    {
+        return [];
+    }
+
+    public function getDialogs(): array
+    {
+        return [];
+    }
+
+    public function getName(): string
+    {
+        return 'DuplicateSetting';
+    }
 }